Added formatted errors and decoupled dispatch handling for dispatcher itself
